@@ -1,4 +1,4 @@
-const About: React.FC = () => {
+const OurStory: React.FC = () => {
   return (
     <div className="text-center">
       <h1 className="text-4xl font-bold mb-4">About Us</h1>
@@ -7,4 +7,4 @@ const About: React.FC = () => {
   );
 };
 
-export default About;
+export default OurStory;
