@@ -8,8 +8,8 @@ export default defineConfig({
     alias: {
       "@components": path.resolve(__dirname, "src/components"),
       "@types": path.resolve(__dirname, "src/types"),
-      "@utils": path.resolve(__dirname, "src/utils"),
-      "@styles": path.resolve(__dirname, "src/styles"),
+      "@assets": path.resolve(__dirname, "src/assets"),
+      "@pages": path.resolve(__dirname, "src/pages"),
     },
   },
 });
